ATR-830: moved RemoveCommonAcumaticaNamespacePrefixes to a more suitable helper class
